requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.2
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-grappelli==3.0.8
django-jazzmin==2.6.0
django-jet==1.0.8
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-suit==0.2.28
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
pycparser==2.21
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
zipp==3.17.0
